pyproject.toml
[project]
name = "terratheme"
version = "0.1.0"
description = "Multi-source-color theme generator for Terra DE"
readme = "README.md"
authors = [
    { name = "Trevin Jones", email = "devtrev00@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "scikit-learn>=1.8.0",
]

[project.scripts]
terratheme = "terratheme.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
